apiconnector metric score with longs

+++ apiconnector/src/main/java/org/openml/apiconnector/models/MetricScore.java
@@ -26,7 +26,7 @@ import org.openml.apiconnector.algorithms.MathHelper;
 public class MetricScore {
 	
 	private Double score = null;
-	private int nr_of_instances = 0;
+	private long nr_of_instances = 0;
 	private Double[] array = null;
 	private double[][] confusion_matrix = null;
 	
@@ -36,7 +36,7 @@ public class MetricScore {
 	 * 
 	 * @param score
 	 */
-	public MetricScore( Double score, int nr_of_instances ) {
+	public MetricScore( Double score, long nr_of_instances ) {
 		this.score = score;
 		this.nr_of_instances = nr_of_instances;
 	}
@@ -47,7 +47,7 @@ public class MetricScore {
 	 * 
 	 * @param array - 
 	 */
-	public MetricScore( Double[] array, int nr_of_instances ) {
+	public MetricScore( Double[] array, long nr_of_instances ) {
 		this.array = array;
 		this.nr_of_instances = nr_of_instances;
 	}
@@ -60,7 +60,7 @@ public class MetricScore {
 	 * @param score 
 	 * @param array
 	 */
-	public MetricScore( Double score, Double[] array, int nr_of_instances ) {
+	public MetricScore( Double score, Double[] array, long nr_of_instances ) {
 		this.score = score;
 		this.array = array;
 		this.nr_of_instances = nr_of_instances;
@@ -94,7 +94,7 @@ public class MetricScore {
 	/**
 	 * @return Returns the number of instances this score was calculated over. 
 	 */
-	public int getNrOfInstances() {
+	public long getNrOfInstances() {
 		return nr_of_instances;
 	}
